move code into getPdbxUnobsOrZeroOccResidues
- explicitely refer to mmcif enum types
